+++ core/list_formula_data.php
@@ -11,23 +11,11 @@ while($cats_res = mysqli_fetch_array($cats_q)){
     $cats[] = $cats_res;
 }
 
-if(empty(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM ingredients")))){
-	$response['Error'] = (string)'<div class="alert alert-info alert-dismissible"><strong>INFO: </strong> no ingredients yet, click <a href="?do=ingredients">here</a> to add.</div>';    
-	header('Content-Type: application/json; charset=utf-8');
-	echo json_encode($response);
-	return;
-}
-if(empty(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM formulasMetaData")))){
-	$response['Error'] = (string)'<div class="alert alert-info alert-dismissible"><strong>INFO: </strong> no formulas yet, click <a href="#" data-toggle="modal" data-target="#add_formula">here</a> to add.</div>';    
-	header('Content-Type: application/json; charset=utf-8');
-	echo json_encode($response);
-	return;
-}
 
 if($_GET['filter'] && $_GET['profile'] || $_GET['sex']){
 	$f = "WHERE profile = '".$_GET['profile']."' OR sex = '".$_GET['sex']."'";
 }
-$formulas = mysqli_query($conn, "SELECT id,fid,name,product_name,isProtected,profile,sex,created,catClass FROM formulasMetaData $f");
+$formulas = mysqli_query($conn, "SELECT id,fid,name,product_name,isProtected,profile,sex,created,catClass,isMade,madeOn FROM formulasMetaData $f");
 
 while ($allFormulas = mysqli_fetch_array($formulas)){
 	    $formula[] = $allFormulas;
@@ -43,6 +31,8 @@ foreach ($formula as $formula) {
 	$r['created'] = (string)$formula['created'];
 	$r['catClass'] = (string)$formula['catClass']?:'N/A';
 	$r['ingredients'] = (int)countElement("formulas WHERE fid = '".$formula['fid']."'",$conn)?:'0';
+	$r['isMade'] = (int)$formula['isMade']?:0;
+	$r['madeOn'] = (string)$formula['madeOn']?:'N/A';
 
 	$response['data'][] = $r;