+++ pages/upload.php
@@ -8,7 +8,95 @@ require_once(__ROOT__.'/inc/settings.php');
 require_once(__ROOT__.'/func/fixIFRACas.php');
 require_once(__ROOT__.'/func/formatBytes.php');
 
+if($_GET['type'] == 'bottle' && $_GET['name']){
+	$name = base64_decode($_GET['name']);
+	$ml = $_GET['size'];
+	$price = $_GET['price'];
+	$height = $_GET['height'];
+	$width = $_GET['width'];
+	$diameter = $_GET['diameter'];
+	$supplier = base64_decode($_GET['supplier']);
+	$supplier_link = base64_decode($_GET['supplier_link']);
+	$notes = base64_decode($_GET['notes']);
+	
+	if(isset($_FILES['pic_file']['name'])){
+      $file_name = $_FILES['pic_file']['name'];
+      $file_size = $_FILES['pic_file']['size'];
+      $file_tmp = $_FILES['pic_file']['tmp_name'];
+      $file_type = $_FILES['pic_file']['type'];
+      $file_ext = strtolower(end(explode('.',$_FILES['pic_file']['name'])));
+	  
+	  if (file_exists('../'.$uploads_path.'bottles/') === FALSE) {
+    	mkdir('../'.$uploads_path.'bottles/', 0740, true);
+	  }
 
+	  $ext = explode(', ', $allowed_ext);
+	  
+      if(in_array($file_ext,$ext)=== false){
+		 echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>File upload error: </strong>Extension not allowed, please choose a '.$allowed_ext.' file.</div>';
+      	return;
+	  }
+	  if($file_size > $max_filesize){
+		 echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>File upload error: </strong>File size must not exceed '.formatBytes($max_filesize).'</div>';
+		 return;
+      }
+	 if(mysqli_num_rows(mysqli_query($conn, "SELECT name FROM bottles WHERE name = '$name'"))){
+		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error: </strong>'.$name.' already exists!</div>';
+	  	return;
+	  }
+      if(move_uploaded_file($file_tmp,'../'.$uploads_path.'bottles/'.base64_encode($file_name))){
+		$photo = $uploads_path.'bottles/'.base64_encode($file_name);
+		
+			if(mysqli_query($conn, "INSERT INTO bottles (name, ml, price, height, width, diameter, supplier, supplier_link, notes, photo) VALUES ('$name', '$ml', '$price', '$height', '$width', '$diameter', '$supplier', '$supplier_link', '$notes', '$photo')") ){
+				echo '<div class="alert alert-success alert-dismissible"><a href="?do=bottles" class="close" data-dismiss="alert" aria-label="close">x</a><strong>'.$name.'</strong> added!</div>';
+			}else{
+				echo '<div class="alert alert-danger alert-dismissible"><a href="?do=bottles" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error:</strong> Failed to add '.$name.' - '.mysqli_error($conn).'</div>';
+			}
+		}
+	  }
+	return;
+}
+
+if($_GET['type'] == 'lid' && $_GET['style']){
+	
+	$style = base64_decode($_GET['style']);
+	$color = base64_decode($_GET['color']);
+	$price = base64_decode($_GET['price']);
+	$supplier = base64_decode($_GET['supplier']);
+	$supplier_link = base64_decode($_GET['supplier_link']);
+
+
+	if(isset($_FILES['pic_file']['name'])){
+      $file_name = $_FILES['pic_file']['name'];
+      $file_size = $_FILES['pic_file']['size'];
+      $file_tmp = $_FILES['pic_file']['tmp_name'];
+      $file_type = $_FILES['pic_file']['type'];
+      $file_ext = strtolower(end(explode('.',$_FILES['pic_file']['name'])));
+	  
+	  if (file_exists('../'.$uploads_path.'lids/') === FALSE) {
+    	mkdir('../'.$uploads_path.'lids/', 0740, true);
+	  }
+
+	  $ext = explode(', ', $allowed_ext);
+	  
+      if(in_array($file_ext,$ext)=== false){
+		 echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>File upload error: </strong>Extension not allowed, please choose a '.$allowed_ext.' file.</div>';
+      }elseif($file_size > $max_filesize){
+		 echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>File upload error: </strong>File size must not exceed '.formatBytes($max_filesize).'</div>';
+      }else{
+	  
+         if(move_uploaded_file($file_tmp,'../'.$uploads_path.'lids/'.base64_encode($file_name))){
+		 	$photo = $uploads_path.'lids/'.base64_encode($file_name);
+			if(mysqli_query($conn, "INSERT INTO lids (style, colour, price, supplier, supplier_link, photo) VALUES ('$style', '$color', '$price', '$supplier', '$supplier_link', '$photo')")){
+		 		echo '<div class="alert alert-success alert-dismissible"><a href="?do=lids" class="close" data-dismiss="alert" aria-label="close">x</a><strong>'.$style.'</strong> added!</div>';
+			}else{
+				echo '<div class="alert alert-danger alert-dismissible"><a href="?do=lids" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error:</strong> Failed to add '.$style.' - '.mysqli_error($conn).'</div>';
+			}
+		 }
+	  }
+   }
+	return;
+}
 
 if($_GET['type'] && $_GET['id']){